+++ dev-packages/e2e-tests/test-applications/create-next-app/tests/server-transactions.test.ts
@@ -40,7 +40,6 @@ test('Sends server-side transactions to Sentry', async ({ baseURL }) => {
       spans: [
         {
           data: {
-            'otel.kind': 'INTERNAL',
             'sentry.origin': 'manual',
           },
           description: 'test-span',