+++ packages/opentelemetry/test/trace.test.ts
@@ -325,6 +325,16 @@ describe('trace', () => {
       expect(getActiveSpan()).toBe(undefined);
     });
 
+    it('allows to pass parentSpan=null', () => {
+      startSpan({ name: 'GET users/[id' }, () => {
+        startSpan({ name: 'child', parentSpan: null }, span => {
+          // Due to the way we propagate the scope in OTEL,
+          // the parent_span_id is not actually undefined here, but comes from the propagation context
+          expect(spanToJSON(span).parent_span_id).toBe(getCurrentScope().getPropagationContext().spanId);
+        });
+      });
+    });
+
     it('allows to force a transaction with forceTransaction=true', async () => {
       const client = getClient()!;
       const transactionEvents: Event[] = [];
@@ -379,7 +389,6 @@ describe('trace', () => {
           'sentry.source': 'custom',
           'sentry.sample_rate': 1,
           'sentry.origin': 'manual',
-          'otel.kind': 'INTERNAL',
         },
         span_id: expect.any(String),
         trace_id: expect.any(String),
@@ -403,7 +412,6 @@ describe('trace', () => {
         data: {
           'sentry.source': 'custom',
           'sentry.origin': 'manual',
-          'otel.kind': 'INTERNAL',
           'sentry.sample_rate': 1,
         },
         parent_span_id: innerParentSpanId,
@@ -579,6 +587,17 @@ describe('trace', () => {
       expect(getActiveSpan()).toBe(undefined);
     });
 
+    it('allows to pass parentSpan=null', () => {
+      startSpan({ name: 'outer' }, () => {
+        const span = startInactiveSpan({ name: 'test span', parentSpan: null });
+
+        // Due to the way we propagate the scope in OTEL,
+        // the parent_span_id is not actually undefined here, but comes from the propagation context
+        expect(spanToJSON(span).parent_span_id).toBe(getCurrentScope().getPropagationContext().spanId);
+        span.end();
+      });
+    });
+
     it('allows to force a transaction with forceTransaction=true', async () => {
       const client = getClient()!;
       const transactionEvents: Event[] = [];
@@ -630,7 +649,6 @@ describe('trace', () => {
           'sentry.source': 'custom',
           'sentry.sample_rate': 1,
           'sentry.origin': 'manual',
-          'otel.kind': 'INTERNAL',
         },
         span_id: expect.any(String),
         trace_id: expect.any(String),
@@ -654,7 +672,6 @@ describe('trace', () => {
         data: {
           'sentry.source': 'custom',
           'sentry.origin': 'manual',
-          'otel.kind': 'INTERNAL',
           'sentry.sample_rate': 1,
         },
         parent_span_id: innerParentSpanId,
@@ -860,6 +877,17 @@ describe('trace', () => {
       expect(getActiveSpan()).toBe(undefined);
     });
 
+    it('allows to pass parentSpan=null', () => {
+      startSpan({ name: 'outer' }, () => {
+        startSpanManual({ name: 'GET users/[id]', parentSpan: null }, span => {
+          // Due to the way we propagate the scope in OTEL,
+          // the parent_span_id is not actually undefined here, but comes from the propagation context
+          expect(spanToJSON(span).parent_span_id).toBe(getCurrentScope().getPropagationContext().spanId);
+          span.end();
+        });
+      });
+    });
+
     it('allows to force a transaction with forceTransaction=true', async () => {
       const client = getClient()!;
       const transactionEvents: Event[] = [];
@@ -918,7 +946,6 @@ describe('trace', () => {
           'sentry.source': 'custom',
           'sentry.sample_rate': 1,
           'sentry.origin': 'manual',
-          'otel.kind': 'INTERNAL',
         },
         span_id: expect.any(String),
         trace_id: expect.any(String),
@@ -942,7 +969,6 @@ describe('trace', () => {
         data: {
           'sentry.source': 'custom',
           'sentry.origin': 'manual',
-          'otel.kind': 'INTERNAL',
           'sentry.sample_rate': 1,
         },
         parent_span_id: innerParentSpanId,