+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 
 - "You miss 100 percent of the chances you don't take. — Wayne Gretzky" — Michael Scott
 
+Work in this release was contributed by @GitSquared. Thank you for your contribution!
+
 ## 8.17.0
 
 - feat: Upgrade OTEL deps (#12809)