+++ CHANGELOG.md
@@ -15,8 +15,7 @@
 The `enableTracing` option has been deprecated and will be removed in the next major version. We recommend removing it
 in favor of the `tracesSampleRate` and `tracesSampler` options. If you want to enable performance monitoring, please set
 the `tracesSampleRate` to a sample rate of your choice, or provide a sampling function as `tracesSampler` option
-instead. If you wan't to disable performance monitoring, remove the `tracesSampler` option and set the
-`tracesSampleRate` to `0` or `undefined`.
+instead. If you wan't to disable performance monitoring, remove the `tracesSampler` and `tracesSampleRate` options.
 
 Work in this release was contributed by @GitSquared. Thank you for your contribution!